+++ src/main/java/com/xiaozhi/websocket/llm/factory/LlmServiceFactory.java
@@ -1,10 +1,11 @@
 package com.xiaozhi.websocket.llm.factory;
 
 import com.xiaozhi.websocket.llm.api.LlmService;
-import com.xiaozhi.websocket.llm.providers.ollama.OllamaService;
-import com.xiaozhi.websocket.llm.providers.openai.OpenAiService;
-import com.xiaozhi.websocket.llm.providers.qwen.QwenService;
-import com.xiaozhi.websocket.llm.providers.spark.SparkService;
+import com.xiaozhi.websocket.llm.providers.OllamaService;
+import com.xiaozhi.websocket.llm.providers.OpenAiService;
+import com.xiaozhi.websocket.llm.providers.QwenService;
+import com.xiaozhi.websocket.llm.providers.SparkService;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -14,19 +15,20 @@ import org.slf4j.LoggerFactory;
  */
 public class LlmServiceFactory {
     private static final Logger logger = LoggerFactory.getLogger(LlmServiceFactory.class);
-    
+
     /**
      * 创建LLM服务
      * 
      * @param provider 提供商
      * @param endpoint API端点
-     * @param apiKey API密钥
-     * @param model 模型名称
+     * @param apiKey   API密钥
+     * @param model    模型名称
      * @return LLM服务
      */
-    public static LlmService createLlmService(String provider, String endpoint, String appId, String apiKey, String apiSecret, String model) {
+    public static LlmService createLlmService(String provider, String endpoint, String appId, String apiKey,
+            String apiSecret, String model) {
         provider = provider.toLowerCase();
-        
+
         switch (provider) {
             case "openai":
                 return new OpenAiService(endpoint, appId, apiKey, apiSecret, model);
